SWM-116 refactor: viewModel 과 Repository 역할 분리
